Testing reducing rate to min on Starlink change
Reduces the bandwidth to the minimum for 4 2 second periods every minute, based on when it seems like Starlink sometimes switches satellites (XX:12, XX:27, XX:42, XX:57). The idea is that fewer packets in flight during these times may lead to less bufferbloat.

This is a hacky work in progress, for testing purposes only.